Доработка сервиса печати.
