+++ Convertors/ConvertDocx.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Sample7.Convertors
+{
+    public static partial class ConvertTo
+    {
+        public static byte[] ConvertDocxToPdf(byte[] inputFileBytes)
+        {
+            //специфика конвертации docx в pdf
+            Console.WriteLine(inputFileBytes);
+            throw new NotImplementedException();
+        }
+    }
+}
