+++ Enums/ImageFormat.cs
@@ -1,0 +1,10 @@
+﻿namespace Sample7.Enums
+{
+    public enum ImageFormat
+    {
+        Jpeg,
+        Png,
+        Tiff,
+        Bmp
+    }
+}
