+++ Enums/PrintServiceSupportedFileType.cs
@@ -1,0 +1,13 @@
+﻿namespace Sample7.Enums
+{
+    public enum PrintServiceSupportedFileType
+    {
+        Unknown = 0,
+        ImageFile,
+        TextFile,
+        CalcFile,
+        DocumFile,
+        HtmlFile,
+        PdfFile
+    }
+}
